# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactPublic": {
        "description": "One artifact in ``GET /runs/:id/artifacts`` (docs/API.md \u00a73.5).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "run_step_id": {
            "title": "Run Step Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "run_step_id",
          "kind",
          "size_bytes",
          "mime_type",
          "created_at"
        ],
        "title": "ArtifactPublic",
        "type": "object"
      },
      "ArtifactSignedUrl": {
        "description": "``GET /runs/:id/artifacts/:artifactId`` \u2014 presigned download URL.",
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "scheme": {
            "title": "Scheme",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "url",
          "kind",
          "scheme",
          "expiresAt"
        ],
        "title": "ArtifactSignedUrl",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMeta": {
        "description": "Pagination metadata: the next cursor (``null`` when exhausted) + page size.",
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "limit"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "Page_RunListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[RunListItem]",
        "type": "object"
      },
...
      "RunDetail": {
        "description": "Detail for ``GET /runs/:id`` \u2014 adds the computed summary.",
        "properties": {
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "env": {
            "title": "Env",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary": {
            "$ref": "#/components/schemas/RunSummary"
          },
          "tier_at_runtime": {
            "$ref": "#/components/schemas/Tier"
          },
          "trigger": {
            "$ref": "#/components/schemas/RunTrigger"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "public_id",
          "project_id",
          "name",
          "env",
          "trigger",
          "status",
          "tier_at_runtime",
          "created_at",
          "updated_at",
          "summary"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "RunLogPage": {
        "description": "A cursor-paginated slice of a run's concatenated stdout/stderr text.",
        "properties": {
          "lines": {
            "items": {
              "type": "string"
            },
            "title": "Lines",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "title": "RunLogPage",
        "type": "object"
      },
      "RunNetworkResponse": {
        "description": "``GET /runs/:id/network`` \u2014 bounded network event list (M1b stub).",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NetworkEvent"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "title": "RunNetworkResponse",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "QUEUED",
          "RUNNING",
          "PASS",
          "FAIL",
          "CANCELLED",
          "ERROR"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "RunTrigger": {
        "enum": [
          "MANUAL",
          "SCHEDULED",
          "CI_PUSH",
          "CI_PR",
          "WEBHOOK",
          "AGENT"
        ],
        "title": "RunTrigger",
        "type": "string"
      },
      "RunsSummary": {
        "description": "``GET /runs/summary`` \u2014 counters for the Runs dashboard summary bar.\n\nField aliases are camelCase to match the M1b frontend client.\n``failed`` folds ``FAIL`` + ``ERROR``; ``avg_duration_ms`` is a workspace-wide\nweighted mean across non-null durations.",
        "properties": {
          "active": {
            "description": "Runs currently in RUNNING state",
            "title": "Active",
            "type": "integer"
          },
          "avgDurationMs": {
            "title": "Avgdurationms",
            "type": "integer"
          },
          "failed": {
            "description": "FAIL + ERROR",
            "title": "Failed",
            "type": "integer"
          },
          "passed": {
            "title": "Passed",
            "type": "integer"
          },
          "queued": {
            "title": "Queued",
            "type": "integer"
          },
          "today": {
            "description": "Runs created since 00:00 UTC",
            "title": "Today",
            "type": "integer"
          }
        },
        "required": [
          "active",
          "today",
          "passed",
          "failed",
          "avgDurationMs",
          "queued"
        ],
        "title": "RunsSummary",
        "type": "object"
      },
...
      "Tier": {
        "description": "Capability tier resolved from env.",
        "enum": [
          "ZERO",
          "LOCAL",
          "CLOUD"
        ],
        "title": "Tier",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "suitest_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Suitest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "description": "List a project's runs with filters; 404 when the project is cross-workspace.",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": true,
            "schema": {
              "title": "Projectid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "branch",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Branch"
            }
          },
          {
            "in": "query",
            "name": "env",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Env"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/summary": {
      "get": {
        "description": "Aggregated counters for the Runs dashboard summary bar (docs/API.md \u00a73.5).\n\nCounts are workspace-scoped (joined via ``projects``). ``failed`` folds\n``FAIL`` + ``ERROR`` together to match the Runs UI's binary outcome card.\nStatic endpoint declared BEFORE the dynamic ``/runs/{run_id}`` route below\nso FastAPI's path matcher doesn't try to treat ``summary`` as a run id.",
        "operationId": "get_runs_summary_api_v1_runs_summary_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Runs Summary",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "description": "Return a run with a step-outcome summary; 404 when cross-workspace.",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/artifacts": {
      "get": {
        "description": "List a run's artifacts; 404 when the run is cross-workspace.",
        "operationId": "get_run_artifacts_api_v1_runs__run_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactPublic"
                  },
                  "title": "Response Get Run Artifacts Api V1 Runs  Run Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Run Artifacts",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/artifacts/{artifact_id}": {
      "get": {
        "description": "Return a presigned download URL for one artifact; 404 if run/artifact unknown.",
        "operationId": "get_artifact_signed_url_api_v1_runs__run_id__artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactSignedUrl"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Artifact Signed Url",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/logs": {
      "get": {
        "description": "Concatenated stdout/stderr text in step_order, paginated by line offset.",
        "operationId": "get_run_logs_api_v1_runs__run_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLogPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Run Logs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/network": {
      "get": {
        "description": "Network events captured during the run (M1b stub).\n\nValidates run-in-workspace scope (404 cross-workspace) and returns an empty\npage \u2014 HAR-driven event extraction lands with the runner in M1c. Frontend\nNetwork tab already renders the empty state today, so wiring this stub now\nmeans the screen stops 404-ing in real dev.",
        "operationId": "get_run_network_api_v1_runs__run_id__network_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunNetworkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Run Network",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/steps": {
      "get": {
        "description": "Return a run's steps (ordered) with outcomes + case public ids; 404 if cross-ws.",
        "operationId": "get_run_steps_api_v1_runs__run_id__steps_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunStepPublic"
                  },
                  "title": "Response Get Run Steps Api V1 Runs  Run Id  Steps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Run Steps",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}